table contract moved to type_traits file
